Dev : refactoring network utils, adding more utilities
