+++ packages/arb-token-bridge-ui/src/components/common/MainNetworkNotSupported.tsx
@@ -1,12 +1,27 @@
 import { useWallet } from '@arbitrum/use-wallet'
 import { Web3Provider } from '@ethersproject/providers'
-import { ChainId, getNetworkName } from '../../util/networks'
-import { switchChain } from '../../util/NetworkUtils'
+import {
+  ChainId,
+  getNetworkLogo,
+  getNetworkName,
+  switchChain
+} from '../../util/networks'
 import { Button } from './Button'
 
-import EthereumLogo from '../../assets/EthereumLogo.png'
-import ArbitrumOneLogo from '../../assets/ArbitrumOneLogo.svg'
-import ArbitrumNovaLogo from '../../assets/ArbitrumNovaLogo.png'
+// info about the logo and the button class of network list
+const networkButtonsStyles: {
+  [key in ChainId]?: { btnThemeClass: string }
+} = {
+  [ChainId.Mainnet]: {
+    btnThemeClass: 'bg-blue-arbitrum'
+  },
+  [ChainId.ArbitrumOne]: {
+    btnThemeClass: 'bg-blue-arbitrum'
+  },
+  [ChainId.ArbitrumNova]: {
+    btnThemeClass: 'bg-brick-dark'
+  }
+}
 
 export const MainNetworkNotSupported = ({
   supportedNetworks
@@ -14,24 +29,6 @@ export const MainNetworkNotSupported = ({
   supportedNetworks: ChainId[]
 }) => {
   const { provider } = useWallet()
-
-  // info about the logo and the button class of network list
-  const networksStyleMap: {
-    [key in ChainId]?: { img: string; btnThemeClass: string }
-  } = {
-    [ChainId.Mainnet]: {
-      img: EthereumLogo,
-      btnThemeClass: 'bg-blue-arbitrum'
-    },
-    [ChainId.ArbitrumOne]: {
-      img: ArbitrumOneLogo,
-      btnThemeClass: 'bg-blue-arbitrum'
-    },
-    [ChainId.ArbitrumNova]: {
-      img: ArbitrumNovaLogo,
-      btnThemeClass: 'bg-brick-dark'
-    }
-  }
 
   return (
     <div className="flex flex-col items-center space-y-8 py-24 px-12">
@@ -52,14 +49,16 @@ export const MainNetworkNotSupported = ({
           }}
           key={chainId}
           className={`text-md ${
-            networksStyleMap?.[Number(chainId) as ChainId]?.['btnThemeClass']
+            networkButtonsStyles?.[Number(chainId) as ChainId]?.[
+              'btnThemeClass'
+            ]
           } w-6/12 min-w-max py-3`}
           aria-label={`Switch to ${getNetworkName(Number(chainId))}`}
         >
           <div className="flex flex-row items-center justify-center space-x-3">
             <div className="flex h-8 w-8 items-center justify-center">
               <img
-                src={networksStyleMap[Number(chainId) as ChainId]?.['img']}
+                src={getNetworkLogo(Number(chainId))}
                 alt={`${getNetworkName(Number(chainId))} logo`}
                 className="max-w-8 max-h-8"
               />
